microservices/llm-qa-module/requirements.txt
# LLMQAModule - Service de Questions/Réponses avec LLM
# Port: 8004

# FastAPI and Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LLM and AI
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai>=1.6.1
sentence-transformers==2.2.2
torch>=2.0.0


# Vector Store
faiss-cpu==1.7.4
chromadb==0.4.22

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Messaging
pika==1.3.2

# Utils
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.1.0
httpx>=0.27.0,<0.28.0
aiohttp==3.9.1
numpy>=1.24.0
tiktoken==0.5.2
